Removes the remaining references to Fl_String.
Still to do: rename fl_filename... to fl_filename..._str
and introduce into core library.

+++ fluid/code.cxx
@@ -798,7 +798,7 @@ int Fd_Code_Writer::write_code(const char *s, const char *t, bool to_codeview) {
   // Remember the last code file location for MergeBack
   if (s && g_project.write_mergeback_data && !to_codeview) {
     std::string proj_filename = g_project.projectfile_path() + g_project.projectfile_name();
-    int i, n = proj_filename.size();
+    int i, n = (int)proj_filename.size();
     for (i=0; i<n; i++) if (proj_filename[i]=='\\') proj_filename[i] = '/';
     Fl_Preferences build_records(Fl_Preferences::USER_L, "fltk.org", "fluid-build");
     Fl_Preferences path(build_records, proj_filename.c_str());
